+++ go.mod
@@ -1,13 +1,13 @@
 module github.com/gofiber/fiber
 
-go 1.13
+go 1.11
 
 require (
-	github.com/CloudyKit/fastprinter v0.0.0-20200109182630-33d98a066a53 // indirect
-	github.com/CloudyKit/jet v2.1.2+incompatible
+	github.com/Joker/jade v1.0.0
 	github.com/aymerick/raymond v2.0.2+incompatible
 	github.com/cbroglie/mustache v1.0.1
 	github.com/eknkc/amber v0.0.0-20171010120322-cdade1c07385
+	github.com/fasthttp/websocket v1.4.1
 	github.com/gorilla/schema v1.1.0
 	github.com/json-iterator/go v1.1.9
 	github.com/valyala/fasthttp v1.9.0
