+++ router.go
@@ -1,164 +1,247 @@
-// 🚀 Fiber is an Express.js inspired web framework written in Go with 💖
-// 📌 Please open an issue if you got suggestions or found a bug!
-// 🖥 Links: https://github.com/gofiber/fiber, https://fiber.wiki
-
-// 🦸 Not all heroes wear capes, thank you to some amazing people
-// 💖 @valyala, @erikdubbelboer, @savsgio, @julienschmidt, @koddr
-
 package fiber
 
 import (
 	"fmt"
 	"log"
-	"path/filepath"
+	"reflect"
 	"regexp"
 	"strings"
-	"sync"
 
+	websocket "github.com/fasthttp/websocket"
 	fasthttp "github.com/valyala/fasthttp"
 )
 
-// Ctx is the context that contains everything
-type Ctx struct {
-	route    *Route
-	next     bool
-	error    error
-	params   *[]string
-	values   []string
-	Fasthttp *fasthttp.RequestCtx
-}
-
 // Route struct
 type Route struct {
-	// HTTP method in uppercase, can be a * for Use() & All()
+	// HTTP method in uppercase, can be a * for "Use" & "All" routes
 	Method string
 	// Stores the original path
 	Path string
-	// Bool that defines if the route is a Use() middleware
-	Midware bool
-	// wildcard bool is for routes without a path, * and /*
-	Wildcard bool
-	// Stores compiled regex special routes :params, *wildcards, optionals?
+	// Prefix is for ending wildcards or middlewares
+	Prefix string
+	// Stores regex for :params & :optionals?
 	Regex *regexp.Regexp
-	// Store params if special routes :params, *wildcards, optionals?
+	// Stores params keys for :params & :optionals?
 	Params []string
-	// Callback function for specific route
-	Handler func(*Ctx)
+	// Callback function for context
+	HandlerCtx func(*Ctx)
+	// Callback function for websockets
+	HandlerConn func(*Conn)
 }
 
-// Ctx pool
-var poolCtx = sync.Pool{
-	New: func() interface{} {
-		return new(Ctx)
-	},
+func (app *App) registerStatic(grpPrefix string, args ...string) {
+	// // Initiate variables
+	// var gzip = true
+	// var root string
+	// var prefix string
+	// var method = "GET"
+	// var path = groupPrefix
+	// // Parse provided path
+	// if len(args) == 1 {
+	// 	root = args[0]
+	// } else if len(args) > 1 {
+	// 	root = args[1]
+	// 	// Non empty paths must start with a "/" or "*"
+	// 	if len(path) > 0 && path[0] != '/' && path[0] != '*' {
+	// 		path = "/" + path
+	// 	}
+	// 	// When grouping, always remove single slash
+	// 	if len(groupPrefix) > 0 && path == "/" {
+	// 		path = ""
+	// 	}
+	// 	// Prepent group prefix if exist
+	// 	path = groupPrefix + path
+	// 	// Clean path by removing double "//" => "/"
+	// 	path = strings.Replace(path, "//", "/", -1)
+	// }
+	// // Set path to "*" if path is "" or "/*"
+	// if path == "/*" {
+	// 	path = "*"
+	// }
+	//
+	// // Set prefix for paths containing only a wildcard
+	// if len(path) > 1 && strings.Contains(path, "*") {
+	// 	// Store prefix for matching routes beggining with...
+	// 	prefix = strings.Split(path, "*")[0]
+	// }
+	// // Get file(s) from root
+	// files, _, err := getFiles(root)
+	// if err != nil {
+	// 	log.Fatal("Static: ", err)
+	// }
+	// // ./static/compiled => static/compiled
+	// mount := filepath.Clean(root)
+	// // Loop over all files
+	// for _, file := range files {
+	// 	// Ignore the .gzipped files by fasthttp
+	// 	if strings.Contains(file, ".fasthttp.gz") {
+	// 		continue
+	// 	}
+	// 	// Time to create a fake path for the route match
+	// 	// static/index.html => /index.html
+	// 	fakePath := filepath.Join(path, strings.Replace(file, mount, "", 1))
+	// 	// for windows: static\index.html => /index.html
+	// 	fakePath = filepath.ToSlash(fakePath)
+	// 	// Store file path to use in ctx handler
+	// 	filePath := file
+	// 	// If the file is an index.html, bind the prefix to index.html directly
+	// 	if filepath.Base(filePath) == "index.html" || filepath.Base(filePath) == "index.htm" {
+	// 		app.routes = append(app.routes, &Route{
+	// 			Method: method,
+	// 			Path:   prefix,
+	// 			Prefix: prefix,
+	// 			HandlerCtx: func(c *Ctx) {
+	// 				c.SendFile(filePath, gzip)
+	// 			},
+	// 		})
+	// 	}
+	// 	app.routes = append(app.routes, &Route{
+	// 		Method: method,
+	// 		Path:   fakePath,
+	// 		Prefix: prefix,
+	// 		HandlerCtx: func(c *Ctx) {
+	// 			c.SendFile(filePath, gzip)
+	// 		},
+	// 	})
+	// }
 }
-
-// Get new Ctx from pool
-func acquireCtx(fctx *fasthttp.RequestCtx) *Ctx {
-	ctx := poolCtx.Get().(*Ctx)
-	ctx.Fasthttp = fctx
-	return ctx
-}
-
-// Return Context to pool
-func releaseCtx(ctx *Ctx) {
-	ctx.route = nil
-	ctx.next = false
-	ctx.error = nil
-	ctx.params = nil
-	ctx.values = nil
-	ctx.Fasthttp = nil
-	poolCtx.Put(ctx)
-}
-
-func (grp *Group) register(method string, args ...interface{}) {
-	path := grp.path
-	var handler func(*Ctx)
-	if len(args) == 1 {
-		handler = args[0].(func(*Ctx))
-	} else if len(args) > 1 {
-		path = path + args[0].(string)
-		handler = args[1].(func(*Ctx))
-		if path[0] != '/' && path[0] != '*' {
-			path = "/" + path
+func (app *App) register(method, grpPrefix string, args ...interface{}) {
+	// Set variables
+	var path = "*"
+	var prefix string
+	var middleware = method == "USE"
+	var handlersCtx []func(*Ctx)
+	var handlersConn []func(*Conn)
+	for i := 0; i < len(args); i++ {
+		switch arg := args[i].(type) {
+		case string:
+			path = arg
+		case func(*Ctx):
+			handlersCtx = append(handlersCtx, arg)
+		case func(*Conn):
+			handlersConn = append(handlersConn, arg)
+		default:
+			log.Fatalf("Invalid argument type: %v", reflect.TypeOf(arg))
 		}
+	}
+	// A non wildcard path must start with a '/'
+	if path != "*" && len(path) > 0 && path[0] != '/' {
+		path = "/" + path
+	}
+	// Prepend group prefix
+	if len(grpPrefix) > 0 {
+		// `/v1`+`/` => `/v1`+``
+		if path == "/" {
+			path = grpPrefix
+		} else {
+			path = grpPrefix + path
+		}
+		// Remove duplicate slashes `//`
 		path = strings.Replace(path, "//", "/", -1)
-		path = filepath.Clean(path)
-		path = filepath.ToSlash(path)
 	}
-	grp.app.register(method, path, handler)
-}
-
-// Function to add a route correctly
-func (app *Application) register(method string, args ...interface{}) {
-	// Set if method is Use() midware
-	var midware = method == "USE"
-
-	// Match any method
-	if method == "ALL" || midware {
-		method = "*"
-	}
-
-	// Prepare possible variables
-	var path string        // We could have a path/prefix
-	var handler func(*Ctx) // We could have a ctx handler
-
-	// Only 1 argument, so no path/prefix
-	if len(args) == 1 {
-		handler = args[0].(func(*Ctx))
-	} else if len(args) > 1 {
-		path = args[0].(string)
-		handler = args[1].(func(*Ctx))
-		if path == "" || path[0] != '/' && path[0] != '*' {
-			path = "/" + path
-		}
-	}
-
-	if midware && strings.Contains(path, "/:") {
-		log.Fatal("Router: You cannot use :params in Use()")
-	}
-
-	// If Use() path == "/", match anything aka *
-	if midware && path == "/" {
+	// Empty or '/*' path equals "match anything"
+	// TODO fix * for paths with grpprefix
+	if path == "" || path == "/*" {
 		path = "*"
 	}
-
-	// If the route needs to match any path
-	if path == "" || path == "*" || path == "/*" {
-		app.routes = append(app.routes, &Route{method, path, midware, true, nil, nil, handler})
+	if method == "ALL" || middleware {
+		method = "*"
+	}
+	// Routes are case insensitive by default
+	if !app.Settings.CaseSensitive {
+		path = strings.ToLower(path)
+	}
+	if !app.Settings.StrictRouting && len(path) > 1 {
+		path = strings.TrimRight(path, "/")
+	}
+	// If the route can match anything
+	if path == "*" {
+		for i := range handlersCtx {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, HandlerCtx: handlersCtx[i],
+			})
+		}
+		for i := range handlersConn {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, HandlerConn: handlersConn[i],
+			})
+		}
 		return
 	}
-
-	// Get params from path
+	// Get ':param' & ':optional?' & '*' from path
 	params := getParams(path)
+	// Enable prefix for midware
+	if len(params) == 0 && middleware {
+		prefix = path
+	}
 
-	// If path has no params (simple path), we don't need regex (also for use())
-	if midware || len(params) == 0 {
-		app.routes = append(app.routes, &Route{method, path, midware, false, nil, nil, handler})
+	// If path has no params (simple path)
+	if len(params) == 0 {
+		for i := range handlersCtx {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, Prefix: prefix, HandlerCtx: handlersCtx[i],
+			})
+		}
+		for i := range handlersConn {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, Prefix: prefix, HandlerConn: handlersConn[i],
+			})
+		}
 		return
 	}
 
-	// We have parametes, so we need to compile regex from the path
+	// If path only contains 1 wildcard, we can create a prefix
+	// If its a middleware, we also create a prefix
+	if len(params) == 1 && params[0] == "*" {
+		prefix = strings.Split(path, "*")[0]
+		for i := range handlersCtx {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, Prefix: prefix,
+				Params: params, HandlerCtx: handlersCtx[i],
+			})
+		}
+		for i := range handlersConn {
+			app.routes = append(app.routes, &Route{
+				Method: method, Path: path, Prefix: prefix,
+				Params: params, HandlerConn: handlersConn[i],
+			})
+		}
+		return
+	}
+	// We have an :param or :optional? and need to compile a regex struct
 	regex, err := getRegex(path)
 	if err != nil {
-		log.Fatal("Router: Invalid url pattern: " + path)
+		log.Fatal("Router: Invalid path pattern: " + path)
 	}
-
-	// Add regex + params to route
-	app.routes = append(app.routes, &Route{method, path, midware, false, regex, params, handler})
+	// Add route with regex
+	for i := range handlersCtx {
+		app.routes = append(app.routes, &Route{
+			Method: method, Path: path, Regex: regex,
+			Params: params, HandlerCtx: handlersCtx[i],
+		})
+	}
+	for i := range handlersConn {
+		app.routes = append(app.routes, &Route{
+			Method: method, Path: path, Regex: regex,
+			Params: params, HandlerConn: handlersConn[i],
+		})
+	}
 }
-
-// then try to match a route as efficient as possible.
-func (app *Application) handler(fctx *fasthttp.RequestCtx) {
-	found := false
-
+func (app *App) handler(fctx *fasthttp.RequestCtx) {
+	// Use this boolean to perform 404 not found at the end
+	var match = false
 	// get custom context from sync pool
 	ctx := acquireCtx(fctx)
-
-	// get path and method from main context
+	// get path and method
 	path := ctx.Path()
+	if !app.Settings.CaseSensitive {
+		path = strings.ToLower(path)
+	}
+	if !app.Settings.StrictRouting && len(path) > 1 {
+		path = strings.TrimRight(path, "/")
+	}
 	method := ctx.Method()
-
+	// enable recovery
 	if app.recover != nil {
 		defer func() {
 			if r := recover(); r != nil {
@@ -167,48 +250,102 @@ func (app *Application) handler(fctx *fasthttp.RequestCtx) {
 			}
 		}()
 	}
-
 	// loop trough routes
 	for _, route := range app.routes {
-		// Skip route if method is not allowed
+		// fmt.Println(route)
+		// Skip route if method does not match
 		if route.Method != "*" && route.Method != method {
 			continue
 		}
-
-		// First check if we match a wildcard or static path
-		if route.Wildcard || route.Path == path {
-			// if route.wildcard || (route.path == path && route.params == nil) {
-			// If * always set the path to the wildcard parameter
-			if route.Wildcard {
+		// Set route pointer if user wants to call .Route()
+		ctx.route = route
+		// wilcard or exact same path
+		// TODO v2: enable or disable case insensitive match
+		if route.Path == "*" || route.Path == path {
+			// if * always set the path to the wildcard parameter
+			if route.Path == "*" {
 				ctx.params = &[]string{"*"}
-				ctx.values = make([]string, 1)
-				ctx.values[0] = path
+				ctx.values = []string{path}
 			}
-			found = true
-			// Set route pointer if user wants to call .Route()
-			ctx.route = route
-			// Execute handler with context
-			route.Handler(ctx)
+			// ctx.Fasthttp.Request.Header.ConnectionUpgrade()
+			// Websocket request
+			if route.HandlerConn != nil && websocket.FastHTTPIsWebSocketUpgrade(fctx) {
+				// Try to upgrade
+				err := socketUpgrade.Upgrade(ctx.Fasthttp, func(fconn *websocket.Conn) {
+					conn := acquireConn(fconn)
+					defer releaseConn(conn)
+					conn.params = ctx.params
+					conn.values = ctx.values
+					releaseCtx(ctx)
+					route.HandlerConn(conn)
+				})
+				// Upgrading failed
+				if err != nil {
+					panic(err)
+				}
+				return
+			}
+			// No handler for HTTP nor websocket
+			if route.HandlerCtx == nil {
+				continue
+			}
+			// Match found, 404 not needed
+			match = true
+			route.HandlerCtx(ctx)
 			// if next is not set, leave loop and release ctx
 			if !ctx.next {
 				break
+			} else {
+				// reset match to false
+				match = false
 			}
 			// set next to false for next iteration
 			ctx.next = false
 			// continue to go to the next route
 			continue
 		}
-
-		// If route is Use() and path starts with route.path
-		// aka strings.HasPrefix(route.path, path)
-		if route.Midware && strings.HasPrefix(path, route.Path) {
-			found = true
+		if route.Prefix != "" && strings.HasPrefix(path, route.Prefix) {
 			ctx.route = route
-			route.Handler(ctx)
+			if strings.Contains(route.Path, "*") {
+				ctx.params = &[]string{"*"}
+				// ctx.values = matches[0][1:len(matches[0])]
+				// parse query source
+				ctx.values = []string{strings.Replace(path, route.Prefix, "", 1)}
+			}
+			// Websocket request
+			if route.HandlerConn != nil {
+				// Try to upgrade
+				err := socketUpgrade.Upgrade(ctx.Fasthttp, func(fconn *websocket.Conn) {
+					conn := acquireConn(fconn)
+					defer releaseConn(conn)
+					conn.params = ctx.params
+					conn.values = ctx.values
+					releaseCtx(ctx)
+					route.HandlerConn(conn)
+				})
+				// Upgrading failed
+				if err != nil {
+					panic(err)
+				}
+				return
+			}
+			// No handler for HTTP nor websocket
+			if route.HandlerCtx == nil {
+				continue
+			}
+			// Match found, 404 not needed
+			match = true
+			route.HandlerCtx(ctx)
+			// if next is not set, leave loop and release ctx
 			if !ctx.next {
 				break
+			} else {
+				// reset match to false
+				match = false
 			}
+			// set next to false for next iteration
 			ctx.next = false
+			// continue to go to the next route
 			continue
 		}
 
@@ -228,34 +365,46 @@ func (app *Application) handler(fctx *fasthttp.RequestCtx) {
 			// If we have matches, add params and values to context
 			if len(matches) > 0 && len(matches[0]) > 1 {
 				ctx.params = &route.Params
-				// ctx.values = make([]string, len(*ctx.params))
 				ctx.values = matches[0][1:len(matches[0])]
 			}
 		}
-
-		found = true
-
-		// Set route pointer if user wants to call .Route()
-		ctx.route = route
-
-		// Execute handler with context
-		route.Handler(ctx)
-
+		// Websocket route
+		if route.HandlerConn != nil {
+			// Try to upgrade
+			err := socketUpgrade.Upgrade(ctx.Fasthttp, func(fconn *websocket.Conn) {
+				conn := acquireConn(fconn)
+				conn.params = ctx.params
+				conn.values = ctx.values
+				releaseCtx(ctx)
+				defer releaseConn(conn)
+				route.HandlerConn(conn)
+			})
+			// Upgrading failed
+			if err != nil {
+				panic(err)
+			}
+			return
+		}
+		// No handler for HTTP nor websocket
+		if route.HandlerCtx == nil {
+			continue
+		}
+		// Match found, 404 not needed
+		match = true
+		route.HandlerCtx(ctx)
 		// if next is not set, leave loop and release ctx
 		if !ctx.next {
 			break
+		} else {
+			// reset match to false
+			match = false
 		}
-
 		// set next to false for next iteration
 		ctx.next = false
 	}
-
-	// No routes found
-	if !found {
-		// Custom 404 handler?
+	// No match, send default 404
+	if !match {
 		ctx.SendStatus(404)
 	}
-
-	// release context back into sync pool
 	releaseCtx(ctx)
 }
