--- router_test.go
@@ -1,10 +1,0 @@
-package fiber
-
-import "testing"
-
-func BenchmarkFib10(b *testing.B) {
-	// run the Fib function b.N times
-	// for n := 0; n < b.N; n++ {
-	// 	Fib(10)
-	// }
-}
